Adding references to relevant issues around feature_columns
