+++ python/ml4ir/applications/ranking/tests/test_ranking_model.py
@@ -67,7 +67,7 @@ class RankingModelTest(RankingTestBase):
 
         # Test model training on CSV data
         data_dir = os.path.join(self.root_data_dir, "csv")
-        feature_config_path = os.path.join(self.root_data_dir, "csv", self.feature_config_fname)
+        feature_config_path = os.path.join(self.root_data_dir, "config", self.feature_config_fname)
 
         csv_loss, csv_mrr = self.run_default_pipeline(
             data_dir=data_dir, data_format="csv", feature_config_path=feature_config_path
@@ -79,9 +79,7 @@ class RankingModelTest(RankingTestBase):
 
         # Test model training on TFRecord SequenceExample data
         data_dir = os.path.join(self.root_data_dir, "tfrecord")
-        feature_config_path = os.path.join(
-            self.root_data_dir, "tfrecord", self.feature_config_fname
-        )
+        feature_config_path = os.path.join(self.root_data_dir, "config", self.feature_config_fname)
 
         tfrecord_loss, tfrecord_mrr = self.run_default_pipeline(
             data_dir=data_dir, data_format="tfrecord", feature_config_path=feature_config_path
