+++ python/ml4ir/base/model/relevance_model.py
@@ -87,6 +87,10 @@ class RelevanceModel:
             )
 
             # Compile model
+            """
+            NOTE:
+            Related Github issue: https://github.com/tensorflow/probability/issues/519
+            """
             self.model.compile(
                 optimizer=optimizer,
                 loss=loss_fn,
